添加会员卡兑换提示信息

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/wxapp/card/vo/CardExchangTipVo.java
@@ -17,10 +17,10 @@ public class CardExchangTipVo {
 	/**
 	 * 周期内可兑换商品数量
 	 */
-	private Byte periodNumber;
+	private Integer periodNumber;
 	
 	/**
 	 * 周期内已经兑换的数量
 	 */
-	private Byte periodExchangGoodsNum;
+	private Integer periodExchangGoodsNum;
 }
